Initial Commit
This is a little program that knows all of the keyword for C# and you can type one in and learn about it. It uses a class called keywords which maintains a dictionary of keyword:definition pairs and using that we can search and find the defined keyword which have definitions that I looked up and made. This program is really just to help me learn C# I may add more to it later?